# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ZENOVA Platform API",
    "description": "Research-Grade Mental-Health Wellbeing and Conversational Support Platform API",
    "contact": {
      "name": "ZENOVA Engineering & Clinical Team",
      "url": "https://zenova.ai/",
      "email": "support@zenova.ai"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/datasets/{name}": {
      "get": {
        "tags": [
          "Dataset Management"
        ],
        "summary": "Get Dataset",
        "description": "Retrieve detailed metadata, splits, and statistics for a specific dataset.",
        "operationId": "get_dataset_api_v1_datasets__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/datasets/validate": {
      "post": {
        "tags": [
          "Dataset Management"
        ],
        "summary": "Validate Dataset",
        "description": "Audit a registered dataset's file checksums and schema integrity.",
        "operationId": "validate_dataset_api_v1_datasets_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ValidateDatasetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/context/build": {
      "post": {
        "tags": [
          "Context Engine"
        ],
        "summary": "Build Context Endpoint",
        "description": "Normalize heterogeneous module outputs into the unified 9-block MultimodalContext.",
        "operationId": "build_context_endpoint_api_v1_context_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContextBuildRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MultimodalContext"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/context/anonymize": {
      "post": {
        "tags": [
          "Context Engine"
        ],
        "summary": "Anonymize Context Endpoint",
        "description": "Apply PII scrubbing and ID pseudonymization to a context payload.",
        "operationId": "anonymize_context_endpoint_api_v1_context_anonymize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnonymizeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/context/{session_id}/latest": {
      "get": {
        "tags": [
          "Context Engine"
        ],
        "summary": "Get Latest Context Snapshot",
        "description": "Retrieve the latest recorded MultimodalContext snapshot for a given session.",
        "operationId": "get_latest_context_snapshot_api_v1_context__session_id__latest_get",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/context/{session_id}/feedback": {
      "post": {
        "tags": [
          "Context Engine"
        ],
        "summary": "Record Feedback Endpoint",
        "description": "Record explicit user feedback on a turn and link it to the session history.",
        "operationId": "record_feedback_endpoint_api_v1_context__session_id__feedback_post",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserFeedback"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/user/data/{user_id}": {
      "delete": {
        "tags": [
          "User Application"
        ],
        "summary": "Purge User Data",
        "description": "Permanently delete all personal user records (GDPR Right to be Forgotten).",
        "operationId": "purge_user_data_api_v1_user_data__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataPurgeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnonymizeRequest": {
        "properties": {
          "context": {
            "additionalProperties": true,
            "type": "object",
            "title": "Context"
          },
          "privacy_level": {
            "$ref": "#/components/schemas/zenova__schemas__context__PrivacyLevel",
            "default": "ANONYMIZED"
          }
        },
        "type": "object",
        "required": [
          "context"
        ],
        "title": "AnonymizeRequest"
      },
...
      "BaselineResult": {
        "properties": {
          "is_placeholder": {
            "type": "boolean",
            "title": "Is Placeholder",
            "default": false
          },
          "module_version": {
            "type": "string",
            "title": "Module Version",
            "default": "1.0.0"
          },
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "Establishment status: insufficient_data, provisional_baseline, or established_baseline",
            "default": "insufficient_data"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 0.0
          },
          "total_observations": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Observations",
            "default": 0
          },
          "z_score_valence": {
            "type": "number",
            "title": "Z Score Valence",
            "default": 0.0
          },
          "is_significant_deviation": {
            "type": "boolean",
            "title": "Is Significant Deviation",
            "default": false
          },
          "deviation_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deviation Notes"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "metric_deviations": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metric Deviations"
          },
          "deviating_features": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Deviating Features"
          },
          "features": {
            "additionalProperties": true,
            "type": "object",
            "title": "Features"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "Statistical deviation from personal baseline; observational and non-diagnostic."
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "user_id"
        ],
        "title": "BaselineResult",
        "description": "Longitudinal baseline deviation comparison."
      },
...
      "ContextBuildRequest": {
        "properties": {
          "user_input": {
            "$ref": "#/components/schemas/UserInput"
          },
          "emotion": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EmotionResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "symptoms": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SymptomResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "risk": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RiskResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "baseline": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BaselineResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "behavior": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BehavioralResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "voice": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VoiceResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "dialog_stage": {
            "$ref": "#/components/schemas/DialogStage",
            "default": "Exploration"
          },
          "privacy_level": {
            "$ref": "#/components/schemas/zenova__schemas__context__PrivacyLevel",
            "default": "STANDARD"
          }
        },
        "type": "object",
        "required": [
          "user_input"
        ],
        "title": "ContextBuildRequest"
      },
...
      "CrisisCategory": {
        "type": "string",
        "enum": [
          "none",
          "self_harm",
          "suicidal_ideation",
          "imminent_danger",
          "violence",
          "delusional_break"
        ],
        "title": "CrisisCategory"
      },
...
      "DataPurgeResponse": {
        "properties": {
          "user_id": {
            "type": "string",
            "title": "User Id"
          },
          "purged_at": {
            "type": "string",
            "title": "Purged At"
          },
          "deleted_sessions": {
            "type": "integer",
            "title": "Deleted Sessions"
          },
          "deleted_turns": {
            "type": "integer",
            "title": "Deleted Turns"
          },
          "deleted_checkins": {
            "type": "integer",
            "title": "Deleted Checkins"
          },
          "deleted_observations": {
            "type": "integer",
            "title": "Deleted Observations"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "PURGED_SUCCESSFULLY"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": "All personal conversational turns, check-ins, baselines, and observations have been permanently deleted."
          }
        },
        "type": "object",
        "required": [
          "user_id",
          "purged_at",
          "deleted_sessions",
          "deleted_turns",
          "deleted_checkins",
          "deleted_observations"
        ],
        "title": "DataPurgeResponse",
        "description": "Confirmation of complete data deletion under the Right to be Forgotten."
      },
...
      "DialogStage": {
        "type": "string",
        "enum": [
          "Exploration",
          "Comforting",
          "Action"
        ],
        "title": "DialogStage"
      },
...
      "EmotionCategory": {
        "type": "string",
        "enum": [
          "joy",
          "sadness",
          "anger",
          "anxiety",
          "fear",
          "neutral",
          "frustration",
          "hope",
          "grief",
          "guilt",
          "shame"
        ],
        "title": "EmotionCategory"
      },
...
      "EmotionResult": {
        "properties": {
          "is_placeholder": {
            "type": "boolean",
            "title": "Is Placeholder",
            "description": "Flag indicating placeholder vs trained model",
            "default": false
          },
          "module_version": {
            "type": "string",
            "title": "Module Version",
            "default": "1.0.0"
          },
          "primary_emotion": {
            "$ref": "#/components/schemas/EmotionCategory"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "probabilities": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Probabilities"
          },
          "valence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Valence",
            "description": "Pleasantness: -1.0 to 1.0",
            "default": 0.0
          },
          "arousal": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Arousal",
            "description": "Activation energy: -1.0 to 1.0",
            "default": 0.0
          },
          "dominance": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Dominance",
            "description": "Perceived control: -1.0 to 1.0",
            "default": 0.0
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "primary_emotion",
          "confidence"
        ],
        "title": "EmotionResult",
        "description": "Emotion analysis output schema."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModalityType": {
        "type": "string",
        "enum": [
          "text",
          "voice",
          "multimodal"
        ],
        "title": "ModalityType"
      },
...
      "MultimodalContext": {
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/ConversationContextBlock"
          },
          "emotion": {
            "$ref": "#/components/schemas/EmotionContextBlock"
          },
          "symptoms": {
            "$ref": "#/components/schemas/SymptomsContextBlock"
          },
          "risk": {
            "$ref": "#/components/schemas/RiskContextBlock"
          },
          "baseline": {
            "$ref": "#/components/schemas/BaselineContextBlock"
          },
          "behavior": {
            "$ref": "#/components/schemas/BehaviorContextBlock"
          },
          "voice": {
            "$ref": "#/components/schemas/VoiceContextBlock"
          },
          "history": {
            "$ref": "#/components/schemas/HistoryContextBlock"
          },
          "metadata": {
            "$ref": "#/components/schemas/MetadataContextBlock"
          }
        },
        "type": "object",
        "required": [
          "conversation",
          "emotion",
          "symptoms",
          "risk",
          "baseline",
          "behavior",
          "voice",
          "history",
          "metadata"
        ],
        "title": "MultimodalContext",
        "description": "The master normalized context object for ZENOVA Step 9.\n\nContains exactly the 9 required blocks:\n- conversation\n- emotion\n- symptoms\n- risk\n- baseline\n- behavior\n- voice\n- history\n- metadata"
      },
...
      "RiskLevel": {
        "type": "string",
        "enum": [
          "low",
          "moderate",
          "high",
          "critical"
        ],
        "title": "RiskLevel"
      },
      "RiskResult": {
        "properties": {
          "is_placeholder": {
            "type": "boolean",
            "title": "Is Placeholder",
            "default": false
          },
          "module_version": {
            "type": "string",
            "title": "Module Version",
            "default": "1.0.0"
          },
          "risk_level": {
            "$ref": "#/components/schemas/RiskLevel",
            "default": "low"
          },
          "crisis_category": {
            "$ref": "#/components/schemas/CrisisCategory",
            "default": "none"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "requires_immediate_escalation": {
            "type": "boolean",
            "title": "Requires Immediate Escalation",
            "default": false
          },
          "requires_escalation": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requires Escalation"
          },
          "trigger_cues": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Trigger Cues"
          },
          "escalation_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Escalation Action"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "confidence"
        ],
        "title": "RiskResult",
        "description": "Crisis and high-risk safety assessment result."
      },
...
      "SymptomResult": {
        "properties": {
          "is_placeholder": {
            "type": "boolean",
            "title": "Is Placeholder",
            "default": false
          },
          "module_version": {
            "type": "string",
            "title": "Module Version",
            "default": "1.0.0"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/SymptomSignal"
            },
            "type": "array",
            "title": "Signals"
          },
          "aggregate_severity": {
            "$ref": "#/components/schemas/SymptomSeverity",
            "default": "none"
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "Informational signals only; not a formal psychiatric diagnosis."
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "title": "SymptomResult",
        "description": "Symptom and signal analysis result."
      },
      "SymptomSeverity": {
        "type": "string",
        "enum": [
          "none",
          "subclinical",
          "mild",
          "moderate",
          "severe"
        ],
        "title": "SymptomSeverity"
      },
...
      "UserFeedback": {
        "properties": {
          "turn_id": {
            "type": "integer",
            "title": "Turn Id"
          },
          "rating": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 5.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Rating",
            "description": "1-5 star user rating"
          },
          "is_helpful": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Helpful"
          },
          "feedback_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Text"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "turn_id"
        ],
        "title": "UserFeedback",
        "description": "Explicit feedback submitted by the user on a specific turn."
      },
      "UserInput": {
        "properties": {
          "session_id": {
            "type": "string",
            "title": "Session Id",
            "description": "Unique active session identifier"
          },
          "user_id": {
            "type": "string",
            "title": "User Id",
            "description": "Anonymized user identifier"
          },
          "text": {
            "type": "string",
            "minLength": 1,
            "title": "Text",
            "description": "Raw textual input from user"
          },
          "modality": {
            "$ref": "#/components/schemas/ModalityType",
            "default": "text"
          },
          "audio_features": {
            "anyOf": [
              {
                "additionalProperties": {
                  "type": "number"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Features",
            "description": "Acoustic features if voice"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "session_id",
          "user_id",
          "text"
        ],
        "title": "UserInput",
        "description": "Inbound user turn with multimodal support and session metadata."
      },
...
      "ValidateDatasetRequest": {
        "properties": {
          "dataset_name": {
            "type": "string",
            "title": "Dataset Name"
          }
        },
        "type": "object",
        "required": [
          "dataset_name"
        ],
        "title": "ValidateDatasetRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "zenova__schemas__context__PrivacyLevel": {
        "type": "string",
        "enum": [
          "STANDARD",
          "ANONYMIZED",
          "EPHEMERAL"
        ],
        "title": "PrivacyLevel"
      },
...
    }
  },
...
}